refactor(riscv): localize load store test support
