+++ extensions/riscv/circuit/src/load/doubleword/tests.rs
@@ -1,20 +1,74 @@
-#[cfg(feature = "cuda")]
-use openvm_circuit::arch::testing::TestBuilder;
-#[cfg(feature = "cuda")]
-use openvm_instructions::LocalOpcode;
+use std::borrow::BorrowMut;
 
-use crate::test_utils::memory::{
-    create_doubleword_harness, create_seeded_rng, load_memory_config, load_write_data,
-    rv64_bytes_to_u16_block, set_and_execute_load, VmChipTestBuilder, LOADD,
+#[cfg(feature = "cuda")]
+use openvm_circuit::arch::testing::{
+    default_var_range_checker_bus, GpuChipTestBuilder, GpuTestChipHarness,
+};
+use openvm_circuit::arch::testing::{TestBuilder, TestChipHarness, VmChipTestBuilder};
+#[cfg(feature = "cuda")]
+use openvm_instructions::riscv::RV64_MEMORY_AS;
+use openvm_instructions::LocalOpcode;
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, LOADD};
+use openvm_stark_backend::{
+    p3_air::BaseAir,
+    p3_field::PrimeCharacteristicRing,
+    p3_matrix::{
+        dense::{DenseMatrix, RowMajorMatrix},
+        Matrix,
+    },
+    utils::disable_debug_builder,
+};
+use openvm_stark_sdk::utils::create_seeded_rng;
+
+use crate::{
+    adapters::{
+        rv64_bytes_to_u16_block, Rv64LoadAdapterAir, Rv64LoadAdapterExecutor, Rv64LoadAdapterFiller,
+    },
+    load::{
+        common::load_write_data, core::LoadCoreCols, LoadDoublewordCoreAir, LoadDoublewordFiller,
+        Rv64LoadDoublewordAir, Rv64LoadDoublewordChip, Rv64LoadDoublewordExecutor,
+        LOAD_DOUBLEWORD_SELECTOR_WIDTH,
+    },
+    test_utils::memory::{load_memory_config, set_and_execute_load, F, MAX_INS_CAPACITY},
 };
 #[cfg(feature = "cuda")]
-use crate::test_utils::memory::{
-    default_var_range_checker_bus, dummy_range_checker, load_gpu_memory_config,
-    transfer_load_records, GpuChipTestBuilder, GpuTestChipHarness, LoadDoublewordCoreAir,
-    LoadDoublewordFiller, Rv64LoadAdapterAir, Rv64LoadAdapterExecutor, Rv64LoadAdapterFiller,
-    Rv64LoadDoublewordAir, Rv64LoadDoublewordChip, Rv64LoadDoublewordChipGpu,
-    Rv64LoadDoublewordExecutor, Rv64LoadStoreOpcode, F, MAX_INS_CAPACITY, RV64_MEMORY_AS,
+use crate::{
+    load::Rv64LoadDoublewordChipGpu,
+    test_utils::memory::{dummy_range_checker, load_gpu_memory_config, transfer_load_records},
 };
+
+type DoublewordHarness = TestChipHarness<
+    F,
+    Rv64LoadDoublewordExecutor,
+    Rv64LoadDoublewordAir,
+    Rv64LoadDoublewordChip<F>,
+>;
+
+fn create_doubleword_harness(tester: &mut VmChipTestBuilder<F>) -> DoublewordHarness {
+    let range_checker = tester.range_checker();
+    let air = Rv64LoadDoublewordAir::new(
+        Rv64LoadAdapterAir::new(
+            tester.memory_bridge(),
+            tester.execution_bridge(),
+            range_checker.bus(),
+            tester.address_bits(),
+        ),
+        LoadDoublewordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
+    );
+    let executor = Rv64LoadDoublewordExecutor::new(
+        Rv64LoadAdapterExecutor::new(tester.address_bits()),
+        Rv64LoadStoreOpcode::CLASS_OFFSET,
+    );
+    let chip = Rv64LoadDoublewordChip::<F>::new(
+        LoadDoublewordFiller::new(
+            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
+            Rv64LoadStoreOpcode::CLASS_OFFSET,
+            range_checker,
+        ),
+        tester.memory_helper(),
+    );
+    DoublewordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
+}
 
 #[test]
 fn rand_load_doubleword_test() {
@@ -46,6 +100,49 @@ fn rand_load_doubleword_test() {
 fn run_loadd_sanity_test() {
     let read_data = rv64_bytes_to_u16_block([138, 45, 202, 76, 131, 74, 186, 29]);
     assert_eq!(load_write_data(LOADD, read_data, 0), read_data);
+}
+
+fn assert_pranked_load_doubleword_fails(
+    prank: impl Fn(&mut LoadCoreCols<F, LOAD_DOUBLEWORD_SELECTOR_WIDTH>),
+) {
+    let mut rng = create_seeded_rng();
+    let mut tester = VmChipTestBuilder::from_config(load_memory_config());
+    let mut harness = create_doubleword_harness(&mut tester);
+    set_and_execute_load(
+        &mut tester,
+        &mut harness.executor,
+        &mut harness.arena,
+        &mut rng,
+        LOADD,
+        None,
+        None,
+        None,
+        None,
+    );
+    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
+    let modify_trace = |trace: &mut DenseMatrix<F>| {
+        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
+        let (_, core_row) = trace_row.split_at_mut(adapter_width);
+        prank(core_row.borrow_mut());
+        *trace = RowMajorMatrix::new(trace_row, trace.width());
+    };
+    disable_debug_builder();
+    tester
+        .build()
+        .load_and_prank_trace(harness, modify_trace)
+        .finalize()
+        .simple_test()
+        .expect_err("pranked doubleword load trace should fail");
+}
+
+#[test]
+fn negative_split_write_data_test() {
+    assert_pranked_load_doubleword_fails(|core| core.read_data[0] += F::ONE);
+}
+
+#[test]
+fn negative_split_opcode_role_test() {
+    assert_pranked_load_doubleword_fails(|core| core.selector[0] += F::ONE);
 }
 
 #[cfg(feature = "cuda")]
