+++ extensions/riscv/circuit/src/load/halfword/tests.rs
@@ -1,20 +1,68 @@
-#[cfg(feature = "cuda")]
-use openvm_circuit::arch::testing::TestBuilder;
-#[cfg(feature = "cuda")]
-use openvm_instructions::LocalOpcode;
+use std::borrow::BorrowMut;
 
-use crate::test_utils::memory::{
-    create_halfword_harness, create_seeded_rng, load_memory_config, load_write_data,
-    rv64_bytes_to_u16_block, set_and_execute_load, VmChipTestBuilder, LOADHU, RV64_MEMORY_AS,
+#[cfg(feature = "cuda")]
+use openvm_circuit::arch::testing::{
+    default_var_range_checker_bus, GpuChipTestBuilder, GpuTestChipHarness,
+};
+use openvm_circuit::arch::testing::{TestBuilder, TestChipHarness, VmChipTestBuilder};
+use openvm_instructions::{riscv::RV64_MEMORY_AS, LocalOpcode};
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, LOADHU};
+use openvm_stark_backend::{
+    p3_air::BaseAir,
+    p3_field::PrimeCharacteristicRing,
+    p3_matrix::{
+        dense::{DenseMatrix, RowMajorMatrix},
+        Matrix,
+    },
+    utils::disable_debug_builder,
+};
+use openvm_stark_sdk::utils::create_seeded_rng;
+
+use crate::{
+    adapters::{
+        rv64_bytes_to_u16_block, Rv64LoadAdapterAir, Rv64LoadAdapterExecutor, Rv64LoadAdapterFiller,
+    },
+    load::{
+        common::load_write_data, core::LoadCoreCols, LoadHalfwordCoreAir, LoadHalfwordFiller,
+        Rv64LoadHalfwordAir, Rv64LoadHalfwordChip, Rv64LoadHalfwordExecutor,
+        LOAD_HALFWORD_SELECTOR_WIDTH,
+    },
+    test_utils::memory::{load_memory_config, set_and_execute_load, F, MAX_INS_CAPACITY},
 };
 #[cfg(feature = "cuda")]
-use crate::test_utils::memory::{
-    default_var_range_checker_bus, dummy_range_checker, load_gpu_memory_config,
-    transfer_load_records, GpuChipTestBuilder, GpuTestChipHarness, LoadHalfwordCoreAir,
-    LoadHalfwordFiller, Rv64LoadAdapterAir, Rv64LoadAdapterExecutor, Rv64LoadAdapterFiller,
-    Rv64LoadHalfwordAir, Rv64LoadHalfwordChip, Rv64LoadHalfwordChipGpu, Rv64LoadHalfwordExecutor,
-    Rv64LoadStoreOpcode, F, MAX_INS_CAPACITY,
+use crate::{
+    load::Rv64LoadHalfwordChipGpu,
+    test_utils::memory::{dummy_range_checker, load_gpu_memory_config, transfer_load_records},
 };
+
+type HalfwordHarness =
+    TestChipHarness<F, Rv64LoadHalfwordExecutor, Rv64LoadHalfwordAir, Rv64LoadHalfwordChip<F>>;
+
+fn create_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> HalfwordHarness {
+    let range_checker = tester.range_checker();
+    let air = Rv64LoadHalfwordAir::new(
+        Rv64LoadAdapterAir::new(
+            tester.memory_bridge(),
+            tester.execution_bridge(),
+            range_checker.bus(),
+            tester.address_bits(),
+        ),
+        LoadHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
+    );
+    let executor = Rv64LoadHalfwordExecutor::new(
+        Rv64LoadAdapterExecutor::new(tester.address_bits()),
+        Rv64LoadStoreOpcode::CLASS_OFFSET,
+    );
+    let chip = Rv64LoadHalfwordChip::<F>::new(
+        LoadHalfwordFiller::new(
+            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
+            Rv64LoadStoreOpcode::CLASS_OFFSET,
+            range_checker,
+        ),
+        tester.memory_helper(),
+    );
+    HalfwordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
+}
 
 #[test]
 fn positive_loadhu_shift6_test() {
@@ -85,6 +133,39 @@ fn run_loadhu_sanity_test() {
         load_write_data(LOADHU, read_data, 6),
         rv64_bytes_to_u16_block([186, 29, 0, 0, 0, 0, 0, 0])
     );
+}
+
+#[test]
+fn negative_split_write_data_test() {
+    let mut rng = create_seeded_rng();
+    let mut tester = VmChipTestBuilder::from_config(load_memory_config());
+    let mut harness = create_halfword_harness(&mut tester);
+    set_and_execute_load(
+        &mut tester,
+        &mut harness.executor,
+        &mut harness.arena,
+        &mut rng,
+        LOADHU,
+        None,
+        None,
+        None,
+        None,
+    );
+    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
+    let modify_trace = |trace: &mut DenseMatrix<F>| {
+        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
+        let (_, core_row) = trace_row.split_at_mut(adapter_width);
+        let core: &mut LoadCoreCols<F, LOAD_HALFWORD_SELECTOR_WIDTH> = core_row.borrow_mut();
+        core.read_data[0] += F::ONE;
+        *trace = RowMajorMatrix::new(trace_row, trace.width());
+    };
+    disable_debug_builder();
+    tester
+        .build()
+        .load_and_prank_trace(harness, modify_trace)
+        .finalize()
+        .simple_test()
+        .expect_err("pranked halfword load trace should fail");
 }
 
 #[cfg(feature = "cuda")]
