+++ extensions/riscv/circuit/src/load_sign_extend/test_utils.rs
@@ -1,195 +1,41 @@
-pub(crate) use std::{array, borrow::BorrowMut, sync::Arc};
+use std::array;
+#[cfg(feature = "cuda")]
+use std::sync::Arc;
 
-pub(crate) use openvm_circuit::{
+use openvm_circuit::{
     arch::{
-        testing::{
-            memory::gen_pointer, TestBuilder, TestChipHarness, VmChipTestBuilder,
-            BITWISE_OP_LOOKUP_BUS,
-        },
+        testing::{memory::gen_pointer, TestBuilder},
         Arena, MemoryConfig, PreflightExecutor,
     },
     system::memory::merkle::public_values::PUBLIC_VALUES_AS,
 };
-pub(crate) use openvm_circuit_primitives::bitwise_op_lookup::{
-    BitwiseOperationLookupAir, BitwiseOperationLookupBus, BitwiseOperationLookupChip,
-    SharedBitwiseOperationLookupChip,
-};
-pub(crate) use openvm_instructions::{
-    instruction::Instruction, riscv::RV64_REGISTER_AS, LocalOpcode,
-};
-pub(crate) use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, *};
-pub(crate) use openvm_stark_backend::{
-    p3_air::BaseAir,
-    p3_field::PrimeCharacteristicRing,
-    p3_matrix::{
-        dense::{DenseMatrix, RowMajorMatrix},
-        Matrix,
-    },
-    utils::disable_debug_builder,
-};
-pub(crate) use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
-pub(crate) use rand::{rngs::StdRng, Rng};
+use openvm_instructions::{instruction::Instruction, riscv::RV64_REGISTER_AS, LocalOpcode};
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, LOADB, LOADH, LOADW};
+use openvm_stark_backend::p3_field::PrimeCharacteristicRing;
+use openvm_stark_sdk::p3_baby_bear::BabyBear;
+use rand::{rngs::StdRng, Rng};
 #[cfg(feature = "cuda")]
-pub(crate) use {
-    super::{
-        Rv64LoadSignExtendByteChipGpu, Rv64LoadSignExtendHalfwordChipGpu,
-        Rv64LoadSignExtendWordChipGpu,
-    },
+use {
     crate::{adapters::Rv64LoadAdapterRecord, load::LoadRecord},
     openvm_circuit::arch::{
-        testing::{
-            default_bitwise_lookup_bus, default_var_range_checker_bus, GpuChipTestBuilder,
-            GpuTestChipHarness,
-        },
+        testing::{default_var_range_checker_bus, GpuTestChipHarness},
         EmptyAdapterCoreLayout,
     },
     openvm_circuit_primitives::var_range::VariableRangeCheckerChip,
 };
 
-pub(crate) use super::{
-    byte::{
-        LoadSignExtendByteCoreAir, LoadSignExtendByteCoreCols, LoadSignExtendByteFiller,
-        Rv64LoadSignExtendByteAir, Rv64LoadSignExtendByteChip, Rv64LoadSignExtendByteExecutor,
-    },
-    halfword::{
-        LoadSignExtendHalfwordCoreAir, LoadSignExtendHalfwordFiller, Rv64LoadSignExtendHalfwordAir,
-        Rv64LoadSignExtendHalfwordChip, Rv64LoadSignExtendHalfwordExecutor,
-        LOAD_SIGN_EXTEND_HALFWORD_SELECTOR_WIDTH,
-    },
-    word::{
-        LoadSignExtendWordCoreAir, LoadSignExtendWordFiller, Rv64LoadSignExtendWordAir,
-        Rv64LoadSignExtendWordChip, Rv64LoadSignExtendWordExecutor,
-        LOAD_SIGN_EXTEND_WORD_SELECTOR_WIDTH,
-    },
-};
-pub(crate) use crate::{
+#[cfg(feature = "cuda")]
+use crate::adapters::Rv64LoadAdapterExecutor;
+use crate::{
     adapters::{
         rv64_bytes_to_u16_block, rv64_bytes_to_u32, rv64_u16_block_to_bytes, sign_extend_imm16,
-        Rv64LoadAdapterAir, Rv64LoadAdapterExecutor, Rv64LoadAdapterFiller, RV64_BYTE_BITS,
     },
-    load_sign_extend::{common::load_sign_extend_write_data, core::LoadSignExtendCoreCols},
+    load_sign_extend::common::load_sign_extend_write_data,
 };
 
 pub(crate) const IMM_BITS: usize = 16;
 pub(crate) const MAX_INS_CAPACITY: usize = 128;
 pub(crate) type F = BabyBear;
-
-pub(crate) type ByteHarness = TestChipHarness<
-    F,
-    Rv64LoadSignExtendByteExecutor,
-    Rv64LoadSignExtendByteAir,
-    Rv64LoadSignExtendByteChip<F>,
->;
-pub(crate) type HalfwordHarness = TestChipHarness<
-    F,
-    Rv64LoadSignExtendHalfwordExecutor,
-    Rv64LoadSignExtendHalfwordAir,
-    Rv64LoadSignExtendHalfwordChip<F>,
->;
-pub(crate) type WordHarness = TestChipHarness<
-    F,
-    Rv64LoadSignExtendWordExecutor,
-    Rv64LoadSignExtendWordAir,
-    Rv64LoadSignExtendWordChip<F>,
->;
-
-pub(crate) fn create_byte_harness(
-    tester: &mut VmChipTestBuilder<F>,
-) -> (
-    ByteHarness,
-    (
-        BitwiseOperationLookupAir<RV64_BYTE_BITS>,
-        SharedBitwiseOperationLookupChip<RV64_BYTE_BITS>,
-    ),
-) {
-    let range_checker = tester.range_checker();
-    let bitwise_bus = BitwiseOperationLookupBus::new(BITWISE_OP_LOOKUP_BUS);
-    let bitwise_chip = Arc::new(BitwiseOperationLookupChip::<RV64_BYTE_BITS>::new(
-        bitwise_bus,
-    ));
-    let air = Rv64LoadSignExtendByteAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadSignExtendByteCoreAir::new(
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip.bus(),
-            range_checker.bus(),
-        ),
-    );
-    let executor = Rv64LoadSignExtendByteExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadSignExtendByteChip::<F>::new(
-        LoadSignExtendByteFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip.clone(),
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    (
-        ByteHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY),
-        (bitwise_chip.air, bitwise_chip),
-    )
-}
-
-pub(crate) fn create_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> HalfwordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64LoadSignExtendHalfwordAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadSignExtendHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadSignExtendHalfwordExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadSignExtendHalfwordChip::<F>::new(
-        LoadSignExtendHalfwordFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    HalfwordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
-
-pub(crate) fn create_word_harness(tester: &mut VmChipTestBuilder<F>) -> WordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64LoadSignExtendWordAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadSignExtendWordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, range_checker.bus()),
-    );
-    let executor = Rv64LoadSignExtendWordExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadSignExtendWordChip::<F>::new(
-        LoadSignExtendWordFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    WordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn set_and_execute<RA: Arena, E: PreflightExecutor<F, RA>>(
@@ -284,100 +130,6 @@ pub(crate) fn memory_config_for() -> MemoryConfig {
     mem_config.addr_spaces[RV64_REGISTER_AS as usize].num_cells = 1 << 29;
     mem_config.addr_spaces[PUBLIC_VALUES_AS as usize].num_cells = 1 << 29;
     mem_config
-}
-pub(crate) fn assert_pranked_byte_fails(prank: impl Fn(&mut LoadSignExtendByteCoreCols<F>)) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let (mut harness, bitwise) = create_byte_harness(&mut tester);
-    set_and_execute(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADB,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .load_periphery(bitwise)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked signed byte load trace should fail");
-}
-
-pub(crate) fn assert_pranked_halfword_fails(
-    prank: impl Fn(&mut LoadSignExtendCoreCols<F, LOAD_SIGN_EXTEND_HALFWORD_SELECTOR_WIDTH>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let mut harness = create_halfword_harness(&mut tester);
-    set_and_execute(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADH,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked signed halfword load trace should fail");
-}
-
-pub(crate) fn assert_pranked_word_fails(
-    prank: impl Fn(&mut LoadSignExtendCoreCols<F, LOAD_SIGN_EXTEND_WORD_SELECTOR_WIDTH>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(memory_config_for());
-    let mut harness = create_word_harness(&mut tester);
-    set_and_execute(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADW,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked signed word load trace should fail");
 }
 
 // ////////////////////////////////////////////////////////////////////////////////////
