+++ extensions/riscv/circuit/src/store/byte/tests.rs
@@ -1,21 +1,96 @@
-#[cfg(feature = "cuda")]
-use openvm_circuit::arch::testing::TestBuilder;
-#[cfg(feature = "cuda")]
-use openvm_instructions::LocalOpcode;
+use std::{borrow::BorrowMut, sync::Arc};
 
-use crate::test_utils::memory::{
-    create_seeded_rng, create_store_byte_harness, rv64_bytes_to_u16_block, set_and_execute_store,
-    store_memory_config, store_write_data, VmChipTestBuilder, STOREB,
+use openvm_circuit::arch::testing::{
+    TestBuilder, TestChipHarness, VmChipTestBuilder, BITWISE_OP_LOOKUP_BUS,
 };
 #[cfg(feature = "cuda")]
-use crate::test_utils::memory::{
-    default_bitwise_lookup_bus, default_var_range_checker_bus, dummy_range_checker,
-    store_gpu_memory_config, transfer_store_records, Arc, BitwiseOperationLookupChip,
-    GpuChipTestBuilder, GpuTestChipHarness, Rv64LoadStoreOpcode, Rv64StoreAdapterAir,
-    Rv64StoreAdapterExecutor, Rv64StoreAdapterFiller, Rv64StoreByteAir, Rv64StoreByteChip,
-    Rv64StoreByteChipGpu, Rv64StoreByteExecutor, StoreByteCoreAir, StoreByteFiller, F,
-    MAX_INS_CAPACITY, PUBLIC_VALUES_AS, RV64_BYTE_BITS, RV64_MEMORY_AS,
+use openvm_circuit::{
+    arch::testing::{
+        default_bitwise_lookup_bus, default_var_range_checker_bus, GpuChipTestBuilder,
+        GpuTestChipHarness,
+    },
+    system::memory::merkle::public_values::PUBLIC_VALUES_AS,
 };
+use openvm_circuit_primitives::bitwise_op_lookup::{
+    BitwiseOperationLookupAir, BitwiseOperationLookupBus, BitwiseOperationLookupChip,
+    SharedBitwiseOperationLookupChip,
+};
+#[cfg(feature = "cuda")]
+use openvm_instructions::riscv::RV64_MEMORY_AS;
+use openvm_instructions::LocalOpcode;
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, STOREB};
+use openvm_stark_backend::{
+    p3_air::BaseAir,
+    p3_field::PrimeCharacteristicRing,
+    p3_matrix::{
+        dense::{DenseMatrix, RowMajorMatrix},
+        Matrix,
+    },
+    utils::disable_debug_builder,
+};
+use openvm_stark_sdk::utils::create_seeded_rng;
+
+use crate::{
+    adapters::{
+        rv64_bytes_to_u16_block, Rv64StoreAdapterAir, Rv64StoreAdapterExecutor,
+        Rv64StoreAdapterFiller, RV64_BYTE_BITS,
+    },
+    store::{
+        common::store_write_data, Rv64StoreByteAir, Rv64StoreByteChip, Rv64StoreByteExecutor,
+        StoreByteCoreAir, StoreByteCoreCols, StoreByteFiller,
+    },
+    test_utils::memory::{set_and_execute_store, store_memory_config, F, MAX_INS_CAPACITY},
+};
+#[cfg(feature = "cuda")]
+use crate::{
+    store::Rv64StoreByteChipGpu,
+    test_utils::memory::{dummy_range_checker, store_gpu_memory_config, transfer_store_records},
+};
+
+type StoreByteHarness =
+    TestChipHarness<F, Rv64StoreByteExecutor, Rv64StoreByteAir, Rv64StoreByteChip<F>>;
+
+fn create_store_byte_harness(
+    tester: &mut VmChipTestBuilder<F>,
+) -> (
+    StoreByteHarness,
+    (
+        BitwiseOperationLookupAir<RV64_BYTE_BITS>,
+        SharedBitwiseOperationLookupChip<RV64_BYTE_BITS>,
+    ),
+) {
+    let range_checker = tester.range_checker();
+    let bitwise_bus = BitwiseOperationLookupBus::new(BITWISE_OP_LOOKUP_BUS);
+    let bitwise_chip = Arc::new(BitwiseOperationLookupChip::<RV64_BYTE_BITS>::new(
+        bitwise_bus,
+    ));
+    let air = Rv64StoreByteAir::new(
+        Rv64StoreAdapterAir::new(
+            tester.memory_bridge(),
+            tester.execution_bridge(),
+            range_checker.bus(),
+            tester.address_bits(),
+        ),
+        StoreByteCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, bitwise_chip.bus()),
+    );
+    let executor = Rv64StoreByteExecutor::new(
+        Rv64StoreAdapterExecutor::new(tester.address_bits()),
+        Rv64LoadStoreOpcode::CLASS_OFFSET,
+    );
+    let chip = Rv64StoreByteChip::<F>::new(
+        StoreByteFiller::new(
+            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
+            Rv64LoadStoreOpcode::CLASS_OFFSET,
+            bitwise_chip.clone(),
+            range_checker,
+        ),
+        tester.memory_helper(),
+    );
+    (
+        StoreByteHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY),
+        (bitwise_chip.air, bitwise_chip),
+    )
+}
 
 #[test]
 fn rand_store_byte_test() {
@@ -80,6 +155,40 @@ fn run_storeb_sanity_test() {
         store_write_data(STOREB, read_data, prev_data, 7),
         rv64_bytes_to_u16_block([199, 83, 243, 12, 90, 121, 64, 221])
     );
+}
+
+#[test]
+fn negative_split_write_data_test() {
+    let mut rng = create_seeded_rng();
+    let mut tester = VmChipTestBuilder::from_config(store_memory_config());
+    let (mut harness, bitwise) = create_store_byte_harness(&mut tester);
+    set_and_execute_store(
+        &mut tester,
+        &mut harness.executor,
+        &mut harness.arena,
+        &mut rng,
+        STOREB,
+        None,
+        None,
+        None,
+        None,
+    );
+    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
+    let modify_trace = |trace: &mut DenseMatrix<F>| {
+        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
+        let (_, core_row) = trace_row.split_at_mut(adapter_width);
+        let core: &mut StoreByteCoreCols<F> = core_row.borrow_mut();
+        core.read_data[0] += F::ONE;
+        *trace = RowMajorMatrix::new(trace_row, trace.width());
+    };
+    disable_debug_builder();
+    tester
+        .build()
+        .load_and_prank_trace(harness, modify_trace)
+        .load_periphery(bitwise)
+        .finalize()
+        .simple_test()
+        .expect_err("pranked byte store trace should fail");
 }
 
 #[cfg(feature = "cuda")]
