+++ extensions/riscv/circuit/src/store/halfword/tests.rs
@@ -1,21 +1,72 @@
-#[cfg(feature = "cuda")]
-use openvm_circuit::arch::testing::TestBuilder;
-#[cfg(feature = "cuda")]
-use openvm_instructions::LocalOpcode;
+use std::borrow::BorrowMut;
 
-use crate::test_utils::memory::{
-    create_seeded_rng, create_store_halfword_harness, rv64_bytes_to_u16_block,
-    set_and_execute_store, store_memory_config, store_write_data, VmChipTestBuilder, STOREH,
+use openvm_circuit::arch::testing::{TestBuilder, TestChipHarness, VmChipTestBuilder};
+#[cfg(feature = "cuda")]
+use openvm_circuit::{
+    arch::testing::{default_var_range_checker_bus, GpuChipTestBuilder, GpuTestChipHarness},
+    system::memory::merkle::public_values::PUBLIC_VALUES_AS,
 };
 #[cfg(feature = "cuda")]
-use crate::test_utils::memory::{
-    default_var_range_checker_bus, dummy_range_checker, store_gpu_memory_config,
-    transfer_store_records, GpuChipTestBuilder, GpuTestChipHarness, Rv64LoadStoreOpcode,
-    Rv64StoreAdapterAir, Rv64StoreAdapterExecutor, Rv64StoreAdapterFiller, Rv64StoreHalfwordAir,
-    Rv64StoreHalfwordChip, Rv64StoreHalfwordChipGpu, Rv64StoreHalfwordExecutor,
-    StoreHalfwordCoreAir, StoreHalfwordFiller, F, MAX_INS_CAPACITY, PUBLIC_VALUES_AS,
-    RV64_MEMORY_AS,
+use openvm_instructions::riscv::RV64_MEMORY_AS;
+use openvm_instructions::LocalOpcode;
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, STOREH};
+use openvm_stark_backend::{
+    p3_air::BaseAir,
+    p3_field::PrimeCharacteristicRing,
+    p3_matrix::{
+        dense::{DenseMatrix, RowMajorMatrix},
+        Matrix,
+    },
+    utils::disable_debug_builder,
 };
+use openvm_stark_sdk::utils::create_seeded_rng;
+
+use crate::{
+    adapters::{
+        rv64_bytes_to_u16_block, Rv64StoreAdapterAir, Rv64StoreAdapterExecutor,
+        Rv64StoreAdapterFiller,
+    },
+    store::{
+        common::store_write_data, core::StoreCoreCols, Rv64StoreHalfwordAir, Rv64StoreHalfwordChip,
+        Rv64StoreHalfwordExecutor, StoreHalfwordCoreAir, StoreHalfwordFiller,
+        STORE_HALFWORD_SELECTOR_WIDTH,
+    },
+    test_utils::memory::{set_and_execute_store, store_memory_config, F, MAX_INS_CAPACITY},
+};
+#[cfg(feature = "cuda")]
+use crate::{
+    store::Rv64StoreHalfwordChipGpu,
+    test_utils::memory::{dummy_range_checker, store_gpu_memory_config, transfer_store_records},
+};
+
+type StoreHalfwordHarness =
+    TestChipHarness<F, Rv64StoreHalfwordExecutor, Rv64StoreHalfwordAir, Rv64StoreHalfwordChip<F>>;
+
+fn create_store_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> StoreHalfwordHarness {
+    let range_checker = tester.range_checker();
+    let air = Rv64StoreHalfwordAir::new(
+        Rv64StoreAdapterAir::new(
+            tester.memory_bridge(),
+            tester.execution_bridge(),
+            range_checker.bus(),
+            tester.address_bits(),
+        ),
+        StoreHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
+    );
+    let executor = Rv64StoreHalfwordExecutor::new(
+        Rv64StoreAdapterExecutor::new(tester.address_bits()),
+        Rv64LoadStoreOpcode::CLASS_OFFSET,
+    );
+    let chip = Rv64StoreHalfwordChip::<F>::new(
+        StoreHalfwordFiller::new(
+            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
+            Rv64LoadStoreOpcode::CLASS_OFFSET,
+            range_checker,
+        ),
+        tester.memory_helper(),
+    );
+    StoreHalfwordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
+}
 
 #[test]
 fn rand_store_halfword_test() {
@@ -63,6 +114,39 @@ fn run_storeh_sanity_test() {
         store_write_data(STOREH, read_data, prev_data, 6),
         rv64_bytes_to_u16_block([144, 56, 175, 92, 90, 121, 250, 123])
     );
+}
+
+#[test]
+fn negative_split_opcode_role_test() {
+    let mut rng = create_seeded_rng();
+    let mut tester = VmChipTestBuilder::from_config(store_memory_config());
+    let mut harness = create_store_halfword_harness(&mut tester);
+    set_and_execute_store(
+        &mut tester,
+        &mut harness.executor,
+        &mut harness.arena,
+        &mut rng,
+        STOREH,
+        None,
+        None,
+        None,
+        None,
+    );
+    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
+    let modify_trace = |trace: &mut DenseMatrix<F>| {
+        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
+        let (_, core_row) = trace_row.split_at_mut(adapter_width);
+        let core: &mut StoreCoreCols<F, STORE_HALFWORD_SELECTOR_WIDTH> = core_row.borrow_mut();
+        core.selector[0] += F::ONE;
+        *trace = RowMajorMatrix::new(trace_row, trace.width());
+    };
+    disable_debug_builder();
+    tester
+        .build()
+        .load_and_prank_trace(harness, modify_trace)
+        .finalize()
+        .simple_test()
+        .expect_err("pranked halfword store trace should fail");
 }
 
 #[cfg(feature = "cuda")]
