+++ extensions/riscv/circuit/src/store/word/tests.rs
@@ -1,20 +1,73 @@
-#[cfg(feature = "cuda")]
-use openvm_circuit::arch::testing::TestBuilder;
-#[cfg(feature = "cuda")]
-use openvm_instructions::LocalOpcode;
+use std::borrow::BorrowMut;
 
-use crate::test_utils::memory::{
-    create_seeded_rng, create_store_word_harness, rv64_bytes_to_u16_block, set_and_execute_store,
-    store_memory_config, store_write_data, VmChipTestBuilder, PUBLIC_VALUES_AS, STOREW,
+#[cfg(feature = "cuda")]
+use openvm_circuit::arch::testing::{
+    default_var_range_checker_bus, GpuChipTestBuilder, GpuTestChipHarness,
+};
+use openvm_circuit::{
+    arch::testing::{TestBuilder, TestChipHarness, VmChipTestBuilder},
+    system::memory::merkle::public_values::PUBLIC_VALUES_AS,
 };
 #[cfg(feature = "cuda")]
-use crate::test_utils::memory::{
-    default_var_range_checker_bus, dummy_range_checker, store_gpu_memory_config,
-    transfer_store_records, GpuChipTestBuilder, GpuTestChipHarness, Rv64LoadStoreOpcode,
-    Rv64StoreAdapterAir, Rv64StoreAdapterExecutor, Rv64StoreAdapterFiller, Rv64StoreWordAir,
-    Rv64StoreWordChip, Rv64StoreWordChipGpu, Rv64StoreWordExecutor, StoreWordCoreAir,
-    StoreWordFiller, F, MAX_INS_CAPACITY, RV64_MEMORY_AS,
+use openvm_instructions::riscv::RV64_MEMORY_AS;
+use openvm_instructions::{LocalOpcode, DEFERRAL_AS};
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, STOREW};
+use openvm_stark_backend::{
+    p3_air::BaseAir,
+    p3_field::PrimeCharacteristicRing,
+    p3_matrix::{
+        dense::{DenseMatrix, RowMajorMatrix},
+        Matrix,
+    },
+    utils::disable_debug_builder,
 };
+use openvm_stark_sdk::utils::create_seeded_rng;
+
+use crate::{
+    adapters::{
+        rv64_bytes_to_u16_block, Rv64StoreAdapterAir, Rv64StoreAdapterCols,
+        Rv64StoreAdapterExecutor, Rv64StoreAdapterFiller,
+    },
+    store::{
+        common::store_write_data, Rv64StoreWordAir, Rv64StoreWordChip, Rv64StoreWordExecutor,
+        StoreWordCoreAir, StoreWordFiller,
+    },
+    test_utils::memory::{set_and_execute_store, store_memory_config, F, MAX_INS_CAPACITY},
+};
+#[cfg(feature = "cuda")]
+use crate::{
+    store::Rv64StoreWordChipGpu,
+    test_utils::memory::{dummy_range_checker, store_gpu_memory_config, transfer_store_records},
+};
+
+type StoreWordHarness =
+    TestChipHarness<F, Rv64StoreWordExecutor, Rv64StoreWordAir, Rv64StoreWordChip<F>>;
+
+fn create_store_word_harness(tester: &mut VmChipTestBuilder<F>) -> StoreWordHarness {
+    let range_checker = tester.range_checker();
+    let air = Rv64StoreWordAir::new(
+        Rv64StoreAdapterAir::new(
+            tester.memory_bridge(),
+            tester.execution_bridge(),
+            range_checker.bus(),
+            tester.address_bits(),
+        ),
+        StoreWordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
+    );
+    let executor = Rv64StoreWordExecutor::new(
+        Rv64StoreAdapterExecutor::new(tester.address_bits()),
+        Rv64LoadStoreOpcode::CLASS_OFFSET,
+    );
+    let chip = Rv64StoreWordChip::<F>::new(
+        StoreWordFiller::new(
+            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
+            Rv64LoadStoreOpcode::CLASS_OFFSET,
+            range_checker,
+        ),
+        tester.memory_helper(),
+    );
+    StoreWordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
+}
 
 #[test]
 fn positive_storew_public_values_test() {
@@ -97,6 +150,39 @@ fn run_storew_sanity_test() {
         store_write_data(STOREW, read_data, prev_data, 4),
         rv64_bytes_to_u16_block([159, 213, 89, 34, 138, 45, 202, 76])
     );
+}
+
+#[test]
+fn negative_split_store_deferral_as_test() {
+    let mut rng = create_seeded_rng();
+    let mut tester = VmChipTestBuilder::from_config(store_memory_config());
+    let mut harness = create_store_word_harness(&mut tester);
+    set_and_execute_store(
+        &mut tester,
+        &mut harness.executor,
+        &mut harness.arena,
+        &mut rng,
+        STOREW,
+        None,
+        None,
+        None,
+        None,
+    );
+    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
+    let modify_trace = |trace: &mut DenseMatrix<F>| {
+        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
+        let (adapter_row, _) = trace_row.split_at_mut(adapter_width);
+        let adapter: &mut Rv64StoreAdapterCols<F> = adapter_row.borrow_mut();
+        adapter.mem_as = F::from_u32(DEFERRAL_AS);
+        *trace = RowMajorMatrix::new(trace_row, trace.width());
+    };
+    disable_debug_builder();
+    tester
+        .build()
+        .load_and_prank_trace(harness, modify_trace)
+        .finalize()
+        .simple_test()
+        .expect_err("pranked store adapter trace should fail");
 }
 
 #[cfg(feature = "cuda")]
