+++ extensions/riscv/circuit/src/test_utils/memory.rs
@@ -1,114 +1,48 @@
-pub(crate) use std::{array, borrow::BorrowMut, sync::Arc};
+use std::array;
+#[cfg(feature = "cuda")]
+use std::sync::Arc;
 
-pub(crate) use openvm_circuit::{
+use openvm_circuit::{
     arch::{
-        testing::{
-            memory::gen_pointer, TestBuilder, TestChipHarness, VmChipTestBuilder,
-            BITWISE_OP_LOOKUP_BUS,
-        },
+        testing::{memory::gen_pointer, TestBuilder},
         Arena, MemoryConfig, PreflightExecutor, BLOCK_FE_WIDTH,
     },
     system::memory::merkle::public_values::PUBLIC_VALUES_AS,
 };
-pub(crate) use openvm_circuit_primitives::bitwise_op_lookup::{
-    BitwiseOperationLookupAir, BitwiseOperationLookupBus, BitwiseOperationLookupChip,
-    SharedBitwiseOperationLookupChip,
-};
-pub(crate) use openvm_instructions::{
+use openvm_instructions::{
     instruction::Instruction,
     riscv::{RV64_MEMORY_AS, RV64_REGISTER_AS},
     LocalOpcode,
 };
-pub(crate) use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{
-    self, LOADB, LOADBU, LOADD, LOADH, LOADHU, LOADW, LOADWU, STOREB, STORED, STOREH, STOREW,
+use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{
+    self, LOADBU, LOADD, LOADHU, LOADWU, STOREB, STORED, STOREH, STOREW,
 };
-pub(crate) use openvm_stark_backend::{
-    p3_air::BaseAir,
-    p3_field::PrimeCharacteristicRing,
-    p3_matrix::{
-        dense::{DenseMatrix, RowMajorMatrix},
-        Matrix,
-    },
-    utils::disable_debug_builder,
-};
-pub(crate) use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
-pub(crate) use rand::{rngs::StdRng, seq::IndexedRandom, Rng};
+use openvm_stark_backend::p3_field::PrimeCharacteristicRing;
+use openvm_stark_sdk::p3_baby_bear::BabyBear;
+use rand::{rngs::StdRng, seq::IndexedRandom, Rng};
 #[cfg(feature = "cuda")]
-pub(crate) use {
+use {
     crate::adapters::{Rv64LoadAdapterRecord, Rv64StoreAdapterRecord},
-    crate::load::{
-        LoadRecord, Rv64LoadByteChipGpu, Rv64LoadDoublewordChipGpu, Rv64LoadHalfwordChipGpu,
-        Rv64LoadWordChipGpu,
-    },
-    crate::store::{
-        Rv64StoreByteChipGpu, Rv64StoreDoublewordChipGpu, Rv64StoreHalfwordChipGpu,
-        Rv64StoreWordChipGpu, StoreRecord,
-    },
+    crate::load::LoadRecord,
+    crate::store::StoreRecord,
     openvm_circuit::arch::{
-        testing::{
-            default_bitwise_lookup_bus, default_var_range_checker_bus, GpuChipTestBuilder,
-            GpuTestChipHarness,
-        },
+        testing::{default_var_range_checker_bus, GpuTestChipHarness},
         EmptyAdapterCoreLayout,
     },
     openvm_circuit_primitives::var_range::VariableRangeCheckerChip,
 };
 
-pub(crate) use crate::{
-    adapters::{
-        rv64_bytes_to_u16_block, rv64_bytes_to_u32, rv64_u16_block_to_bytes, sign_extend_imm16,
-        Rv64LoadAdapterAir, Rv64LoadAdapterExecutor, Rv64LoadAdapterFiller, Rv64StoreAdapterAir,
-        Rv64StoreAdapterCols, Rv64StoreAdapterExecutor, Rv64StoreAdapterFiller, RV64_BYTE_BITS,
-    },
-    load::{
-        common::load_write_data, core::LoadCoreCols, LoadByteCoreAir, LoadByteCoreCols,
-        LoadByteFiller, LoadDoublewordCoreAir, LoadDoublewordFiller, LoadHalfwordCoreAir,
-        LoadHalfwordFiller, LoadWordCoreAir, LoadWordFiller, Rv64LoadByteAir, Rv64LoadByteChip,
-        Rv64LoadByteExecutor, Rv64LoadDoublewordAir, Rv64LoadDoublewordChip,
-        Rv64LoadDoublewordExecutor, Rv64LoadHalfwordAir, Rv64LoadHalfwordChip,
-        Rv64LoadHalfwordExecutor, Rv64LoadWordAir, Rv64LoadWordChip, Rv64LoadWordExecutor,
-        LOAD_DOUBLEWORD_SELECTOR_WIDTH, LOAD_HALFWORD_SELECTOR_WIDTH, LOAD_WORD_SELECTOR_WIDTH,
-    },
-    load_sign_extend::common::load_sign_extend_write_data,
-    store::{
-        common::store_write_data, core::StoreCoreCols, Rv64StoreByteAir, Rv64StoreByteChip,
-        Rv64StoreByteExecutor, Rv64StoreDoublewordAir, Rv64StoreDoublewordChip,
-        Rv64StoreDoublewordExecutor, Rv64StoreHalfwordAir, Rv64StoreHalfwordChip,
-        Rv64StoreHalfwordExecutor, Rv64StoreWordAir, Rv64StoreWordChip, Rv64StoreWordExecutor,
-        StoreByteCoreAir, StoreByteCoreCols, StoreByteFiller, StoreDoublewordCoreAir,
-        StoreDoublewordFiller, StoreHalfwordCoreAir, StoreHalfwordFiller, StoreWordCoreAir,
-        StoreWordFiller, STORE_HALFWORD_SELECTOR_WIDTH,
-    },
+#[cfg(feature = "cuda")]
+use crate::adapters::{Rv64LoadAdapterExecutor, Rv64StoreAdapterExecutor};
+use crate::{
+    adapters::{rv64_bytes_to_u32, rv64_u16_block_to_bytes, sign_extend_imm16},
+    load::common::load_write_data,
+    store::common::store_write_data,
 };
 
 pub(crate) const IMM_BITS: usize = 16;
 pub(crate) const MAX_INS_CAPACITY: usize = 128;
 pub(crate) type F = BabyBear;
-
-pub(crate) type ByteHarness =
-    TestChipHarness<F, Rv64LoadByteExecutor, Rv64LoadByteAir, Rv64LoadByteChip<F>>;
-pub(crate) type HalfwordHarness =
-    TestChipHarness<F, Rv64LoadHalfwordExecutor, Rv64LoadHalfwordAir, Rv64LoadHalfwordChip<F>>;
-pub(crate) type WordHarness =
-    TestChipHarness<F, Rv64LoadWordExecutor, Rv64LoadWordAir, Rv64LoadWordChip<F>>;
-pub(crate) type DoublewordHarness = TestChipHarness<
-    F,
-    Rv64LoadDoublewordExecutor,
-    Rv64LoadDoublewordAir,
-    Rv64LoadDoublewordChip<F>,
->;
-pub(crate) type StoreByteHarness =
-    TestChipHarness<F, Rv64StoreByteExecutor, Rv64StoreByteAir, Rv64StoreByteChip<F>>;
-pub(crate) type StoreHalfwordHarness =
-    TestChipHarness<F, Rv64StoreHalfwordExecutor, Rv64StoreHalfwordAir, Rv64StoreHalfwordChip<F>>;
-pub(crate) type StoreWordHarness =
-    TestChipHarness<F, Rv64StoreWordExecutor, Rv64StoreWordAir, Rv64StoreWordChip<F>>;
-pub(crate) type StoreDoublewordHarness = TestChipHarness<
-    F,
-    Rv64StoreDoublewordExecutor,
-    Rv64StoreDoublewordAir,
-    Rv64StoreDoublewordChip<F>,
->;
 
 struct MemoryAccess {
     a: usize,
@@ -162,250 +96,6 @@ fn random_memory_access(
         rs1,
         shift_amount,
     }
-}
-
-pub(crate) fn create_byte_harness(
-    tester: &mut VmChipTestBuilder<F>,
-) -> (
-    ByteHarness,
-    (
-        BitwiseOperationLookupAir<RV64_BYTE_BITS>,
-        SharedBitwiseOperationLookupChip<RV64_BYTE_BITS>,
-    ),
-) {
-    let range_checker = tester.range_checker();
-    let bitwise_bus = BitwiseOperationLookupBus::new(BITWISE_OP_LOOKUP_BUS);
-    let bitwise_chip = Arc::new(BitwiseOperationLookupChip::<RV64_BYTE_BITS>::new(
-        bitwise_bus,
-    ));
-    let air = Rv64LoadByteAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadByteCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, bitwise_chip.bus()),
-    );
-    let executor = Rv64LoadByteExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadByteChip::<F>::new(
-        LoadByteFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip.clone(),
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    (
-        ByteHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY),
-        (bitwise_chip.air, bitwise_chip),
-    )
-}
-
-pub(crate) fn create_store_byte_harness(
-    tester: &mut VmChipTestBuilder<F>,
-) -> (
-    StoreByteHarness,
-    (
-        BitwiseOperationLookupAir<RV64_BYTE_BITS>,
-        SharedBitwiseOperationLookupChip<RV64_BYTE_BITS>,
-    ),
-) {
-    let range_checker = tester.range_checker();
-    let bitwise_bus = BitwiseOperationLookupBus::new(BITWISE_OP_LOOKUP_BUS);
-    let bitwise_chip = Arc::new(BitwiseOperationLookupChip::<RV64_BYTE_BITS>::new(
-        bitwise_bus,
-    ));
-    let air = Rv64StoreByteAir::new(
-        Rv64StoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        StoreByteCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET, bitwise_chip.bus()),
-    );
-    let executor = Rv64StoreByteExecutor::new(
-        Rv64StoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64StoreByteChip::<F>::new(
-        StoreByteFiller::new(
-            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            bitwise_chip.clone(),
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    (
-        StoreByteHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY),
-        (bitwise_chip.air, bitwise_chip),
-    )
-}
-
-pub(crate) fn create_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> HalfwordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64LoadHalfwordAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
-    );
-    let executor = Rv64LoadHalfwordExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadHalfwordChip::<F>::new(
-        LoadHalfwordFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    HalfwordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
-
-pub(crate) fn create_store_halfword_harness(
-    tester: &mut VmChipTestBuilder<F>,
-) -> StoreHalfwordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64StoreHalfwordAir::new(
-        Rv64StoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        StoreHalfwordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
-    );
-    let executor = Rv64StoreHalfwordExecutor::new(
-        Rv64StoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64StoreHalfwordChip::<F>::new(
-        StoreHalfwordFiller::new(
-            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    StoreHalfwordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
-
-pub(crate) fn create_word_harness(tester: &mut VmChipTestBuilder<F>) -> WordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64LoadWordAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadWordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
-    );
-    let executor = Rv64LoadWordExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadWordChip::<F>::new(
-        LoadWordFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    WordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
-
-pub(crate) fn create_store_word_harness(tester: &mut VmChipTestBuilder<F>) -> StoreWordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64StoreWordAir::new(
-        Rv64StoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        StoreWordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
-    );
-    let executor = Rv64StoreWordExecutor::new(
-        Rv64StoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64StoreWordChip::<F>::new(
-        StoreWordFiller::new(
-            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    StoreWordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
-
-pub(crate) fn create_doubleword_harness(tester: &mut VmChipTestBuilder<F>) -> DoublewordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64LoadDoublewordAir::new(
-        Rv64LoadAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        LoadDoublewordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
-    );
-    let executor = Rv64LoadDoublewordExecutor::new(
-        Rv64LoadAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64LoadDoublewordChip::<F>::new(
-        LoadDoublewordFiller::new(
-            Rv64LoadAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    DoublewordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
-}
-
-pub(crate) fn create_store_doubleword_harness(
-    tester: &mut VmChipTestBuilder<F>,
-) -> StoreDoublewordHarness {
-    let range_checker = tester.range_checker();
-    let air = Rv64StoreDoublewordAir::new(
-        Rv64StoreAdapterAir::new(
-            tester.memory_bridge(),
-            tester.execution_bridge(),
-            range_checker.bus(),
-            tester.address_bits(),
-        ),
-        StoreDoublewordCoreAir::new(Rv64LoadStoreOpcode::CLASS_OFFSET),
-    );
-    let executor = Rv64StoreDoublewordExecutor::new(
-        Rv64StoreAdapterExecutor::new(tester.address_bits()),
-        Rv64LoadStoreOpcode::CLASS_OFFSET,
-    );
-    let chip = Rv64StoreDoublewordChip::<F>::new(
-        StoreDoublewordFiller::new(
-            Rv64StoreAdapterFiller::new(tester.address_bits(), range_checker.clone()),
-            Rv64LoadStoreOpcode::CLASS_OFFSET,
-            range_checker,
-        ),
-        tester.memory_helper(),
-    );
-    StoreDoublewordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -579,235 +269,6 @@ pub(crate) fn store_gpu_memory_config() -> MemoryConfig {
     mem_config.addr_spaces[PUBLIC_VALUES_AS as usize].num_cells = 1 << mem_config.pointer_max_bits;
     mem_config
 }
-pub(crate) fn assert_pranked_load_byte_fails(prank: impl Fn(&mut LoadByteCoreCols<F>)) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(load_memory_config());
-    let (mut harness, bitwise) = create_byte_harness(&mut tester);
-    set_and_execute_load(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADBU,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .load_periphery(bitwise)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked byte memory access trace should fail");
-}
-
-pub(crate) fn assert_pranked_store_byte_fails(prank: impl Fn(&mut StoreByteCoreCols<F>)) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(store_memory_config());
-    let (mut harness, bitwise) = create_store_byte_harness(&mut tester);
-    set_and_execute_store(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        STOREB,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .load_periphery(bitwise)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked byte store trace should fail");
-}
-
-pub(crate) fn assert_pranked_load_halfword_fails(
-    prank: impl Fn(&mut LoadCoreCols<F, { LOAD_HALFWORD_SELECTOR_WIDTH }>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(load_memory_config());
-    let mut harness = create_halfword_harness(&mut tester);
-    set_and_execute_load(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADHU,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked halfword memory access trace should fail");
-}
-
-pub(crate) fn assert_pranked_store_halfword_fails(
-    prank: impl Fn(&mut StoreCoreCols<F, { STORE_HALFWORD_SELECTOR_WIDTH }>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(store_memory_config());
-    let mut harness = create_store_halfword_harness(&mut tester);
-    set_and_execute_store(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        STOREH,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked halfword store trace should fail");
-}
-
-pub(crate) fn assert_pranked_load_word_fails(
-    prank: impl Fn(&mut LoadCoreCols<F, { LOAD_WORD_SELECTOR_WIDTH }>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(load_memory_config());
-    let mut harness = create_word_harness(&mut tester);
-    set_and_execute_load(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADWU,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked word memory access trace should fail");
-}
-
-pub(crate) fn assert_pranked_load_doubleword_fails(
-    prank: impl Fn(&mut LoadCoreCols<F, { LOAD_DOUBLEWORD_SELECTOR_WIDTH }>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(load_memory_config());
-    let mut harness = create_doubleword_harness(&mut tester);
-    set_and_execute_load(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        LOADD,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (_, core_row) = trace_row.split_at_mut(adapter_width);
-        prank(core_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked doubleword memory access trace should fail");
-}
-
-pub(crate) fn assert_pranked_store_word_adapter_fails(
-    prank: impl Fn(&mut Rv64StoreAdapterCols<F>),
-) {
-    let mut rng = create_seeded_rng();
-    let mut tester = VmChipTestBuilder::from_config(store_memory_config());
-    let mut harness = create_store_word_harness(&mut tester);
-    set_and_execute_store(
-        &mut tester,
-        &mut harness.executor,
-        &mut harness.arena,
-        &mut rng,
-        STOREW,
-        None,
-        None,
-        None,
-        None,
-    );
-    let adapter_width = BaseAir::<F>::width(&harness.air.adapter);
-    let modify_trace = |trace: &mut DenseMatrix<F>| {
-        let mut trace_row = trace.row_slice(0).unwrap().to_vec();
-        let (adapter_row, _) = trace_row.split_at_mut(adapter_width);
-        prank(adapter_row.borrow_mut());
-        *trace = RowMajorMatrix::new(trace_row, trace.width());
-    };
-    disable_debug_builder();
-    tester
-        .build()
-        .load_and_prank_trace(harness, modify_trace)
-        .finalize()
-        .simple_test()
-        .expect_err("pranked store adapter trace should fail");
-}
-
 // ////////////////////////////////////////////////////////////////////////////////////
 //  CUDA TESTS
 //
@@ -841,40 +302,4 @@ pub(crate) fn transfer_store_records<G, C, A, E>(harness: &mut GpuTestChipHarnes
             &mut harness.matrix_arena,
             EmptyAdapterCoreLayout::<F, Rv64StoreAdapterExecutor>::new(),
         );
-}
-
-#[cfg(test)]
-mod tests {
-    use openvm_instructions::DEFERRAL_AS;
-    use openvm_stark_backend::p3_field::PrimeCharacteristicRing;
-
-    use super::{
-        assert_pranked_load_byte_fails, assert_pranked_load_doubleword_fails,
-        assert_pranked_load_halfword_fails, assert_pranked_load_word_fails,
-        assert_pranked_store_byte_fails, assert_pranked_store_halfword_fails,
-        assert_pranked_store_word_adapter_fails, F,
-    };
-
-    #[test]
-    fn negative_split_write_data_tests() {
-        assert_pranked_store_byte_fails(|core| core.read_data[0] += F::ONE);
-        assert_pranked_load_halfword_fails(|core| core.read_data[0] += F::ONE);
-        assert_pranked_load_word_fails(|core| core.read_data[0] += F::ONE);
-        assert_pranked_load_doubleword_fails(|core| core.read_data[0] += F::ONE);
-    }
-
-    #[test]
-    fn negative_split_opcode_role_tests() {
-        assert_pranked_load_byte_fails(|core| core.selector[0] += F::ONE);
-        assert_pranked_store_halfword_fails(|core| core.selector[0] += F::ONE);
-        assert_pranked_load_word_fails(|core| core.selector[0] += F::ONE);
-        assert_pranked_load_doubleword_fails(|core| core.selector[0] += F::ONE);
-    }
-
-    #[test]
-    fn negative_split_store_deferral_as_test() {
-        assert_pranked_store_word_adapter_fails(|adapter| {
-            adapter.mem_as = F::from_u32(DEFERRAL_AS)
-        });
-    }
 }
